[compleseus] Devide search functions into 2 intial input variants
Each consult search function now comes with two variants regarding initial
input. They all initialize the input with region content if a region is active,
but differ on whether an empty string or thing-at-point if a region is not
active.

This change tries to mimic the search behaviors configured in helm layers. Thus
the default empty variants are bound to lowercase keys, and the default
thing-at-point variants are bound to uppercase keys.