fix: Symbol’s function definition is void: pupo/before-popwin-create
Commit f610a4860591eb56dd12fa9ee81dfe3fc19146dd fixed the usage of
`define-advice` in the `spacemacs-purpose` layer functions, but not in the
`spacemacs-purpose-popwin` package.

The error did not appear before because the package was already installed in
`.emacs.d/elpa/...` using the code before the change to define-advice. Once it
was reinstalled, the new error started popping up when trying to pop up (sic)
a new window.

This commit replaces `define-advice` by `advice-add` and only advises the
functions once the mode is activated (instead of during package load).