fix: Symbol’s function definition is void: window-purpose/save-dedicated-windows (#16413)
* fix: Symbol’s function definition is void: window-purpose/save-dedicated-windows

Commit 1ce1b0ea21c2e1e68b0037329941c316b4dfef2e introduced a bug due to
incorrect usage of define-advice.

This commit refactors the change by:
- Extracting the advices to named functions spacemacs/window-purpose-{save,restore}-dedicated-windows
- Using advice-add globally instead of a define-advice/advice-add mix

* fix: Symbol’s function definition is void: pupo/before-popwin-create

Commit f610a4860591eb56dd12fa9ee81dfe3fc19146dd fixed the usage of
`define-advice` in the `spacemacs-purpose` layer functions, but not in the
`spacemacs-purpose-popwin` package.

The error did not appear before because the package was already installed in
`.emacs.d/elpa/...` using the code before the change to define-advice. Once it
was reinstalled, the new error started popping up when trying to pop up (sic)
a new window.

This commit replaces `define-advice` by `advice-add` and only advises the
functions once the mode is activated (instead of during package load).