Make undo-tree optional
Add support for the remaining options of `evil-undo-system`, i.e. `undo-redo`
and `undo-fu`. Uses undo-fu-session and vundo to replicate persistence and a
tree view.

I opted for not changing the default or removing undo-tree completely
immediately without a warning, because this could make users lose precious(?)
persisted undo history.

BTW, the package evil-lisp-state has `undo-tree-undo` and `undo-tree-redo`
hardcoded, this should be reported upstream.