[modeline] Remove redundant function calls
The behaviour of `powerline-reset` does not depend on the current buffer, and it
suffices to call them once. The same holds for `powerline-set-selected-window`
if we call `force-mode-line-update` with a non-nil ALL argument to redisplay all
modelines.

This significantly speeds up reloading the configuration with `SPC f e R` in the
presence of many buffers, where `powerline-reset` used to be the
bottleneck. (For me it often took tens of seconds).