[layouts] Fix loading workspaces in case scratch buffer was killed
If a buffer from an eyebrowse window config is not live, restoring this window
configuration will fail silently. Hence eyebrowse replaces killed buffers
with *scratch* before loading window configurations.

However, the scratch buffer itself can be killed. In that case any window
configuration containing it, or containing killed buffers will fail to load.
Thus we need to ensure that the scratch buffer is live when necessary.