[pdf] Improve performance with transient states
Redisplaying pdf pages can be slow. Similar to how it is already disabled for
minibuffer prompts (see the implementation of
`pdf-view-maybe-redisplay-resized-windows`), do not redisplay when entering a
transient state and while it is active. Note that after exiting a transient
state redisplay will be done if necessary.

This makes a difference when `pdf-view-display-size` is `fit-height`,
or when resizing windows using the window transient state.