Fix space-doc-mode error when cl is not loaded (#16594)
* space-doc.el: Reindent and revise plist syntax to avoid weird indentation

* space-doc.el: Don't use deleted function org-font-lock-ensure

This was an internal-to-org-mode alias for font-lock-ensure, which was
added in Emacs 25.1.  Just call the latter directly.

* space-doc.el: Fix calls to obsolete functions

* space-doc.el: Fix error when cl is not loaded

dolist only has an implicit block named "nil" if cl is loaded (since
that adds advice to dolist).  If only cl-lib is loaded, the callsite
must have either cl-block with the dolist, or cl-dolist.  Otherwise,
this error is issued when enabling space-doc-mode:

(no-catch --cl-block-nil-- "~")