[layouts] Fix loading workspaces in case scratch buffer was killed  (#16598)
* [defaults] Refactor `spacemacs/switch-to-scratch-buffer`

In particular, factor out getting/creating the buffer. This is similar to the
function `get-scratch-buffer-create` added in Emacs 29.

* [layouts] Fix loading workspaces in case scratch buffer was killed

If a buffer from an eyebrowse window config is not live, restoring this window
configuration will fail silently. Hence eyebrowse replaces killed buffers
with *scratch* before loading window configurations.

However, the scratch buffer itself can be killed. In that case any window
configuration containing it, or containing killed buffers will fail to load.
Thus we need to ensure that the scratch buffer is live when necessary.