Use rustic-cargo-test-run instead of rustic-cargo-test (#16614)
* Use rustic-cargo-test-run instead of rustic-cargo-test

The command rustic-cargo-test depends on the variable rustic-test-args which gets polluted when running the "current-test" variant. So there is no way to run all the tests after running a single one, without clearing that variable.
Using rustic-cargo-test-run allows to go back to running all tests instead.
We could also keep both, but a meaningful description would be needed for rustic-cargo-test because the behavior is "unexpected"

* minor

* Fix CHANGELOG typo

* [rust] Use the previous keybinding "ta"

* Fix CHANGELOG syntax