core-themes-support.el: Run hooks after enable-theme not load-theme
The intention of spacemacs-post-theme-change-hook is to run after the
theme is changed, not just after a new theme file is loaded.
Spacemacs itself happens to always call load-theme to enable a theme,
but consult-theme (and maybe other packages) call enable-theme rather
than load-theme if a theme has already been loaded.

In any case, switching to advising enable-theme should be fine because
load-theme calls enable-theme anyway.