[core] Correctly initialize location of built-in packages
Previously, if no location was specified, it was set to the default `elpa`.
This did not mean, however, that the packages where installed from ELPA; the
built-in packages were used regardless (they might not even be published on an
ELPA, of course).

Now we correctly set the location to `built-in`, in preparation to making
`:location elpa` take effect for built-in packages, too.