[core] Directly use `load` for loading lock file
Because the lock file is loaded before `dotspacemacs/user-init` is called, and
due to #16871, I would not introduce a condition-case-unless-debug form here. It
would make it difficult to obtain a backtrace in case of an error (in
particular, setting `debug-on-error` in `dotspacemacs/user-init` would not work,
and `--debug-init` does not work, either).