Support orderless annotation and kwd matchers.
Orderless has a few new matchers that are quite useful. The kwd matcher provides
a lot of different ones based on different modes, and the annotation matcher
with `&` allows you to match against all marginalia annotations.

To enable this we need to change the component separators back to the
default (which is a space not preceded by a backslash).

This could be left as a configuration option, but I figured the nicest solution
was to use the default. People can override if they prefer the other behaviour.