[core] Refactor spacemacs/alternate-buffer
Since Emacs 27, there is the option switch-to-prev-buffer-skip which we can
leverage instead of iterating through the window-prev-buffers ourselves. The
only difference between using previous-buffer and manually switching to the
buffer is that previous-buffer respects and sets window-next-buffers. In order
to keep the same behaviour, we need to set the window's next buffers to nil
before and after invoking previous-buffer. (Note that a window's next buffers are
usually a subset of the window's previous buffers, and switching to a buffer with
any command other than previous-buffer or next-buffer will also set the window's
next buffers to nil.)

This simplifies the function a little bit, and more importantly, enables reusing
the predicate spacemacs//alternate-buffer-skip elsewhere without code
duplication.

Also, deprecate the optional window parameter. It serves no purpose, one can
just as well use with-selected-window at call sites. (And calls using the window
argument will not work with the cycling functionality from the next commit
enabled.)