[navigation] Add cycling functionality to SPC TAB and SPC w TAB
Sometimes one finds oneself in the situation where one used SPC TAB to switch to
the last buffer, but it turns out that one actually intended to switch to the
second or third to last buffer. By enabling the option added in this commit, one
can consecutively press TAB in such situations to cycle through the previous
buffers. From the point of view of subsequent commands it is as though one
switched directly to the destination buffer. This uses the excellent
transient-cycles package under the
hood (https://elpa.gnu.org/packages/transient-cycles.html).

The setup needs lexical binding in spacemacs-navigation/packages.el. I have
audited the file and it does not look as if anything depends on dynamic binding.

I considered (and implemented a first version) using hydra/Spacemacs transient
states, by the way, but the implementation using the transient-cycles package is
cleaner.